refactor(input, output): 입출력 모듈을 util 폴더로 분리한다
기존 Components 폴더에 위치하던 입출력 모듈을 util 폴더로 분리한다.
