+++ src/App.js
@@ -1,6 +1,6 @@
 import Game from './components/Game.js';
-import Input from './components/Input.js';
-import Output from './components/Output.js';
+import Input from './utils/io/Input.js';
+import Output from './utils/io/Output.js';
 import validateNames from './utils/validation/validateNames.js';
 import validateRepetitionString from './utils/validation/validateRepetitionString.js';
 import validateTimeComplexity from './utils/validation/validateTimeComplexity.js';
